+++ packages/extractor/src/extractor/mod.js
@@ -1,42 +1,15 @@
-import { IS_NODE_LIKE } from "../util.js";
+import ts from "typescript";
 
-/**
- * @internal
- * Get the file path to the module, with vfs environments in mind.
- * @param {string} specifier
- * @returns {URL}
- */
-function get_module_url(specifier) {
-	if (!IS_NODE_LIKE) {
-		return new URL(`file:///node_modules/${specifier}`); // Refer to virtual file system
-	}
-	if (globalThis.process?.env.VITEST) {
-		// @ts-expect-error FIXME: Ugly workaround for `import.meta.resolve` not working in Vitest: https://github.com/vitest-dev/vitest/issues/6953
-		//eslint-disable-next-line no-undef
-		__vite_ssr_import_meta__.resolve = (path) =>
-			"file://" +
-			globalThis
-				// @ts-expect-error FIXME: 👆
-				.createRequire(import.meta.url)
-				.resolve(path);
-	}
-	return new URL(import.meta.resolve(specifier));
-}
+import { ComponentDocExtractor } from "./component-doc.js";
+import { Compiler } from "../compiler.js";
+import { Options } from "../options.js";
+import { Parser } from "../parser.js";
 
 /**
  * @import { Tag } from "./component-doc.js";
  * @import { Source } from "../util.js";
  * @import { createCacheStorage } from "../cache.js";
  */
-
-import path from "pathe";
-
-import ts from "typescript";
-
-import { Compiler } from "../compiler.js";
-import { Options } from "../options.js";
-import { Parser } from "../parser.js";
-import { ComponentDocExtractor } from "./component-doc.js";
 
 class Extractor {
 	/** @type {Source} */
@@ -210,83 +183,24 @@ class Extractor {
 	 */
 	#create_program() {
 		const program = ts.createProgram({
-			rootNames: [get_module_url("svelte2tsx/svelte-shims-v4.d.ts").pathname, this.compiler.filepath],
-			options: this.#get_ts_options(),
+			rootNames: [this.#options.shims_path, this.#options.tsx_filepath],
+			options: this.#options.get_ts_options(),
 			host: this.#create_host(),
 			oldProgram: this.#cache.program,
 		});
 		return program;
 	}
 
-	/** @returns {ts.CompilerOptions} */
-	#get_ts_options() {
-		const cached = this.#cache.get(this.compiler.filepath)?.options;
-		if (cached) return cached;
-		const options = this.#build_ts_options();
-		this.#cache.set(this.compiler.filepath, { options });
-		return options;
-	}
-
-	/** @type {ts.CompilerOptions} */
-	#forced_ts_options = {
-		allowJs: true,
-		checkJs: true,
-		noEmit: true,
-		skipDefaultLibCheck: true,
-		skipLibCheck: true,
-		sourceMap: false,
-		strict: true,
-	};
-
-	/** @returns {ts.CompilerOptions} */
-	#build_ts_options() {
-		const configpath = this.#find_ts_config_path();
-		if (configpath) {
-			const config_file = ts.readConfigFile(configpath, this.system.readFile);
-			const parsed = ts.parseJsonConfigFileContent(
-				config_file.config,
-				this.system,
-				path.dirname(configpath),
-				undefined,
-				configpath,
-				undefined,
-				[
-					{
-						extension: "svelte",
-						isMixedContent: true,
-						scriptKind: ts.ScriptKind.Deferred,
-					},
-				],
-			);
-			return {
-				...parsed.options,
-				...this.#forced_ts_options,
-			};
-			// TODO: Create a warning when tsconfig.json `compilerOptions.lib` doesn't include DOM
-		}
-		return this.#forced_ts_options;
-	}
-
-	/**
-	 * @returns {string | undefined}
-	 */
-	#find_ts_config_path() {
-		return (
-			ts.findConfigFile(this.compiler.filepath, this.system.fileExists) ||
-			ts.findConfigFile(this.compiler.filepath, this.system.fileExists, "jsconfig.json")
-		);
-	}
-
 	/**
 	 * @returns {ts.CompilerHost}
 	 */
 	#create_host() {
-		const default_host = this.#options.host || ts.createCompilerHost(this.#get_ts_options());
+		const default_host = this.#options.host || ts.createCompilerHost(this.#options.get_ts_options());
 		/** @type {Partial<ts.CompilerHost>} */
 		const overridden_methods = {
 			fileExists: (filepath) => {
 				if (this.#cache.has(filepath)) return true;
-				if (filepath === this.compiler.filepath) return true;
+				if (filepath === this.#options.tsx_filepath) return true;
 				return default_host.fileExists(filepath);
 			},
 			getSourceFile: (filepath, language_version, on_error) => {
@@ -294,10 +208,10 @@ class Extractor {
 				if (cached?.source) return cached.source;
 				/** @type {ts.SourceFile | undefined} */
 				let source;
-				if (filepath === this.compiler.filepath) {
+				if (filepath === this.#options.tsx_filepath) {
 					const content = this.compiler.tsx.code;
 					source = ts.createSourceFile(
-						this.compiler.filepath,
+						this.#options.tsx_filepath,
 						content,
 						language_version,
 						true,
@@ -316,7 +230,7 @@ class Extractor {
 				if (cached?.content) return cached.content;
 				/** @type {string | undefined} */
 				let content;
-				if (filepath === this.compiler.filepath) content = this.compiler.tsx.code;
+				if (filepath === this.#options.tsx_filepath) content = this.compiler.tsx.code;
 				else content = default_host.readFile(filepath);
 				if (content) this.#cache.set(filepath, { content });
 				return content;
@@ -345,13 +259,13 @@ class Extractor {
 	/** @returns {ts.SourceFile} */
 	get #source_file() {
 		if (this.#cached_source_file) return this.#cached_source_file;
-		const from_cache = this.#cache.get(this.compiler.filepath)?.source;
+		const from_cache = this.#cache.get(this.#options.tsx_filepath)?.source;
 		if (from_cache) return from_cache;
-		const from_program = this.#program.getSourceFile(this.compiler.filepath);
+		const from_program = this.#program.getSourceFile(this.#options.tsx_filepath);
 		//O TODO: Document it
 		if (!from_program)
-			throw new Error(`Source file could not be found by TypeScript program: ${this.compiler.filepath}`);
-		this.#cached_source_file = this.#cache.set(this.compiler.filepath, {
+			throw new Error(`Source file could not be found by TypeScript program: ${this.#options.tsx_filepath}`);
+		this.#cached_source_file = this.#cache.set(this.#options.tsx_filepath, {
 			source: from_program,
 		}).source;
 		return from_program;
